src/knowledge_graph/distributed_quantum_graph.py
import numpy as np
from typing import Dict, List, Any, Optional, Tuple, Set
from dataclasses import dataclass, field
import networkx as nx
from concurrent.futures import ThreadPoolExecutor, as_completed
import scipy.sparse as sp
from scipy.linalg import expm
import qiskit
from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister
from qiskit.quantum_info import Statevector

@dataclass
class QuantumNode:
    """
    Represents a quantum-inspired node in the knowledge graph
    """
    id: str
    state_vector: np.ndarray
    superposition_weights: Dict[str, complex]
    entanglement_pairs: List[Tuple[str, float]]
    quantum_properties: Dict[str, Any] = field(default_factory=dict)

@dataclass
class DistributedShard:
    """
    Represents a distributed shard of the knowledge graph
    """
    shard_id: str
    nodes: Dict[str, QuantumNode]
    adjacency_matrix: sp.csr_matrix
    quantum_state: Optional[Statevector] = None
    local_cache: Dict[str, Any] = field(default_factory=dict)

class DistributedQuantumGraph:
    def __init__(
        self,
        num_shards: int = 4,
        quantum_simulation_depth: int = 2
    ):
        """
        Initialize Distributed Quantum Knowledge Graph
        
        Args:
            num_shards: Number of distributed shards
            quantum_simulation_depth: Depth of quantum circuits
        """
        self.num_shards = num_shards
        self.quantum_depth = quantum_simulation_depth
        
        # Initialize distributed shards
        self.shards: Dict[str, DistributedShard] = {
            f"shard_{i}": DistributedShard(
                shard_id=f"shard_{i}",
                nodes={},
                adjacency_matrix=sp.csr_matrix((0, 0))
            )
            for i in range(num_shards)
        }
        
        # Quantum circuit configuration
        self.quantum_config = {
            'num_qubits': 8,
            'measurement_basis': 'computational',
            'entanglement_scheme': 'circular'
        }
        
        # Initialize quantum backend
        self.quantum_backend = qiskit.Aer.get_backend('statevector_simulator')
    
    def add_node(
        self,
        node_id: str,
        state_vector: Optional[np.ndarray] = None,
        properties: Dict[str, Any] = None
    ) -> QuantumNode:
        """
        Add a quantum node to the distributed graph
        
        Args:
            node_id: Node identifier
            state_vector: Initial quantum state vector
            properties: Node properties
        
        Returns:
            Created quantum node
        """
        # Generate quantum state if not provided
        if state_vector is None:
            state_vector = self._generate_quantum_state()
        
        # Create quantum node
        quantum_node = QuantumNode(
            id=node_id,
            state_vector=state_vector,
            superposition_weights=self._compute_superposition_weights(state_vector),
            entanglement_pairs=[],
            quantum_properties=properties or {}
        )
        
        # Assign to shard
        shard_id = self._get_shard_for_node(node_id)
        self.shards[shard_id].nodes[node_id] = quantum_node
        
        # Update shard's quantum state
        self._update_shard_quantum_state(shard_id)
        
        return quantum_node
    
    def add_edge(
        self,
        source_id: str,
        target_id: str,
        weight: float = 1.0,
        entanglement_strength: float = 0.5
    ):
        """
        Add a quantum edge between nodes
        
        Args:
            source_id: Source node ID
            target_id: Target node ID
            weight: Edge weight
            entanglement_strength: Quantum entanglement strength
        """
        # Get source and target shards
        source_shard = self._get_shard_for_node(source_id)
        target_shard = self._get_shard_for_node(target_id)
        
        # Update adjacency matrices
        self._update_adjacency_matrix(source_shard, source_id, target_id, weight)
        
        # Create quantum entanglement
        self._create_entanglement(
            source_id,
            target_id,
            entanglement_strength
        )
    
    def query_subgraph(
        self,
        node_ids: List[str],
        quantum_enabled: bool = True
    ) -> Tuple[Dict[str, QuantumNode], sp.csr_matrix]:
        """
        Query a subgraph using quantum-inspired algorithms
        
        Args:
            node_ids: List of node IDs to query
            quantum_enabled: Whether to use quantum algorithms
        
        Returns:
            Subgraph nodes and adjacency matrix
        """
        # Collect nodes from shards
        subgraph_nodes = {}
        for node_id in node_ids:
            shard_id = self._get_shard_for_node(node_id)
            if node_id in self.shards[shard_id].nodes:
                subgraph_nodes[node_id] = self.shards[shard_id].nodes[node_id]
        
        # Build subgraph adjacency matrix
        adjacency_matrix = self._build_subgraph_matrix(node_ids)
        
        if quantum_enabled:
            # Apply quantum operations
            quantum_circuit = self._create_quantum_circuit(subgraph_nodes)
            quantum_state = self._execute_quantum_circuit(quantum_circuit)
            
            # Update node states based on quantum computation
            self._update_node_states(subgraph_nodes, quantum_state)
        
        return subgraph_nodes, adjacency_matrix
    
    def _generate_quantum_state(self) -> np.ndarray:
        """
        Generate initial quantum state vector
        
        Returns:
            Quantum state vector
        """
        # Create random quantum state
        state = np.random.rand(2 ** self.quantum_config['num_qubits']) + \
                1j * np.random.rand(2 ** self.quantum_config['num_qubits'])
        
        # Normalize
        return state / np.linalg.norm(state)
    
    def _compute_superposition_weights(
        self,
        state_vector: np.ndarray
    ) -> Dict[str, complex]:
        """
        Compute superposition weights from state vector
        
        Args:
            state_vector: Quantum state vector
        
        Returns:
            Dictionary of basis states and weights
        """
        weights = {}
        num_states = len(state_vector)
        
        for i in range(num_states):
            if abs(state_vector[i]) > 1e-10:  # Ignore negligible amplitudes
                basis_state = format(i, f'0{self.quantum_config["num_qubits"]}b')
                weights[basis_state] = state_vector[i]
        
        return weights
    
    def _get_shard_for_node(self, node_id: str) -> str:
        """
        Determine which shard should store a node
        
        Args:
            node_id: Node identifier
        
        Returns:
            Shard identifier
        """
        # Simple hash-based sharding
        shard_index = hash(node_id) % self.num_shards
        return f"shard_{shard_index}"
    
    def _update_shard_quantum_state(self, shard_id: str):
        """
        Update quantum state of a shard
        
        Args:
            shard_id: Shard identifier
        """
        shard = self.shards[shard_id]
        
        # Create quantum circuit for shard
        circuit = QuantumCircuit(self.quantum_config['num_qubits'])
        
        # Apply quantum operations based on node states
        for node in shard.nodes.values():
            self._apply_node_operations(circuit, node)
        
        # Execute circuit
        shard.quantum_state = self._execute_quantum_circuit(circuit)
    
    def _update_adjacency_matrix(
        self,
        shard_id: str,
        source_id: str,
        target_id: str,
        weight: float
    ):
        """
        Update shard's adjacency matrix
        
        Args:
            shard_id: Shard identifier
            source_id: Source node ID
            target_id: Target node ID
            weight: Edge weight
        """
        shard = self.shards[shard_id]
        
        # Get node indices
        source_idx = list(shard.nodes.keys()).index(source_id)
        target_idx = list(shard.nodes.keys()).index(target_id)
        
        # Update sparse matrix
        if shard.adjacency_matrix.shape[0] <= max(source_idx, target_idx):
            # Resize matrix if needed
            new_size = max(source_idx, target_idx) + 1
            new_matrix = sp.csr_matrix((new_size, new_size))
            new_matrix[:shard.adjacency_matrix.shape[0], :shard.adjacency_matrix.shape[1]] = shard.adjacency_matrix
            shard.adjacency_matrix = new_matrix
        
        shard.adjacency_matrix[source_idx, target_idx] = weight
    
    def _create_entanglement(
        self,
        source_id: str,
        target_id: str,
        strength: float
    ):
        """
        Create quantum entanglement between nodes
        
        Args:
            source_id: Source node ID
            target_id: Target node ID
            strength: Entanglement strength
        """
        # Get source and target nodes
        source_shard = self._get_shard_for_node(source_id)
        target_shard = self._get_shard_for_node(target_id)
        
        source_node = self.shards[source_shard].nodes[source_id]
        target_node = self.shards[target_shard].nodes[target_id]
        
        # Add entanglement pairs
        source_node.entanglement_pairs.append((target_id, strength))
        target_node.entanglement_pairs.append((source_id, strength))
    
    def _build_subgraph_matrix(self, node_ids: List[str]) -> sp.csr_matrix:
        """
        Build adjacency matrix for subgraph
        
        Args:
            node_ids: List of node IDs
        
        Returns:
            Sparse adjacency matrix
        """
        n = len(node_ids)
        matrix = sp.lil_matrix((n, n))
        
        # Build matrix from shard data
        for i, source_id in enumerate(node_ids):
            source_shard = self._get_shard_for_node(source_id)
            shard = self.shards[source_shard]
            
            if source_id in shard.nodes:
                source_idx = list(shard.nodes.keys()).index(source_id)
                for j, target_id in enumerate(node_ids):
                    if target_id in shard.nodes:
                        target_idx = list(shard.nodes.keys()).index(target_id)
                        matrix[i, j] = shard.adjacency_matrix[source_idx, target_idx]
        
        return matrix.tocsr()
    
    def _create_quantum_circuit(
        self,
        nodes: Dict[str, QuantumNode]
    ) -> QuantumCircuit:
        """
        Create quantum circuit for node operations
        
        Args:
            nodes: Dictionary of quantum nodes
        
        Returns:
            Quantum circuit
        """
        circuit = QuantumCircuit(self.quantum_config['num_qubits'])
        
        # Apply node operations
        for node in nodes.values():
            self._apply_node_operations(circuit, node)
        
        # Apply entanglement operations
        self._apply_entanglement_operations(circuit, nodes)
        
        return circuit
    
    def _apply_node_operations(
        self,
        circuit: QuantumCircuit,
        node: QuantumNode
    ):
        """
        Apply quantum operations for a node
        
        Args:
            circuit: Quantum circuit
            node: Quantum node
        """
        # Apply single-qubit gates based on node state
        for i, amplitude in enumerate(node.state_vector):
            if abs(amplitude) > 1e-10:
                # Apply rotation gates
                theta = np.angle(amplitude)
                circuit.rx(theta, i % self.quantum_config['num_qubits'])
                circuit.rz(theta, i % self.quantum_config['num_qubits'])
    
    def _apply_entanglement_operations(
        self,
        circuit: QuantumCircuit,
        nodes: Dict[str, QuantumNode]
    ):
        """
        Apply entanglement operations to quantum circuit
        
        Args:
            circuit: Quantum circuit
            nodes: Dictionary of quantum nodes
        """
        # Apply CNOT gates for entangled pairs
        for node_id, node in nodes.items():
            for target_id, strength in node.entanglement_pairs:
                if target_id in nodes:
                    # Apply controlled operations based on entanglement strength
                    source_qubit = hash(node_id) % self.quantum_config['num_qubits']
                    target_qubit = hash(target_id) % self.quantum_config['num_qubits']
                    
                    circuit.cx(source_qubit, target_qubit)
                    circuit.rz(strength * np.pi, target_qubit)
                    circuit.cx(source_qubit, target_qubit)
    
    def _execute_quantum_circuit(
        self,
        circuit: QuantumCircuit
    ) -> Statevector:
        """
        Execute quantum circuit
        
        Args:
            circuit: Quantum circuit to execute
        
        Returns:
            Final quantum state
        """
        # Execute circuit on quantum backend
        job = qiskit.execute(
            circuit,
            self.quantum_backend,
            shots=1
        )
        result = job.result()
        
        return Statevector.from_instruction(circuit)
    
    def _update_node_states(
        self,
        nodes: Dict[str, QuantumNode],
        quantum_state: Statevector
    ):
        """
        Update node states based on quantum computation
        
        Args:
            nodes: Dictionary of quantum nodes
            quantum_state: Final quantum state
        """
        state_vector = quantum_state.data
        
        # Update each node's state vector and superposition weights
        for i, (node_id, node) in enumerate(nodes.items()):
            # Extract relevant part of state vector
            start_idx = i * (2 ** (self.quantum_config['num_qubits'] // len(nodes)))
            end_idx = (i + 1) * (2 ** (self.quantum_config['num_qubits'] // len(nodes)))
            
            node.state_vector = state_vector[start_idx:end_idx]
            node.superposition_weights = self._compute_superposition_weights(node.state_vector)
